Increasing lock timeout for cache. Making tx serialization level explicit. Will not remove active work spaces when creating new one. Adding (failing) concurrency test.

+++ SvnBridgeLibrary/SourceControl/TFSSourceControlProvider.cs
@@ -370,6 +370,8 @@ namespace SvnBridge.SourceControl
 			{
 				if (workspace.Comment != Constants.WorkspaceComment)
 					continue;
+				if (ActivityRepository.Exists(workspace.Name))
+					continue;
 				SourceControlService.DeleteWorkspace(serverUrl, credentials,
 													 workspace.Name);
 			}
